refactor(FE): 재사용을 위해 Github 로그인 로직을 훅에서 모두 처리하도록 변경 (#147)
